moves loadBackend data to maincoomponent
